Extract functions from routes.js
Move TaxonomyData and RelatedContent into modules.
